Move getCaseSensitiveFileName() to compiler sensor and write only one warning per unresolvable file path
