+++ sonar-cxx-plugin/src/main/java/org/sonar/plugins/cxx/compiler/CxxCompilerSensor.java
@@ -30,6 +30,8 @@ import org.sonar.plugins.cxx.utils.CxxUtils;
 
 import java.io.File;
 import java.util.HashSet;
+import java.util.Iterator;
+import java.util.List;
 import java.util.Scanner;
 import java.util.regex.Pattern;
 
@@ -99,6 +101,7 @@ public class CxxCompilerSensor extends CxxReportSensor {
         String id = scanner.match().group(3);
         String msg = scanner.match().group(4);
         // get filename from file system - e.g. VC writes case insensitive file name to html
+        filename = getCaseSensitiveFileName(filename, project.getFileSystem().getSourceDirs());
         CxxUtils.LOG.debug("Scanner-matches file='" + filename + "' line='" + line + "' id='" + id + "' msg=" + msg);
         if (isInputValid(filename, line, id, msg)) {
             if (uniqueIssues.add(filename + line + id + msg))
@@ -124,4 +127,39 @@ public class CxxCompilerSensor extends CxxReportSensor {
       && !StringUtils.isEmpty(id) && !StringUtils.isEmpty(msg);
   }
 
+  /**
+   *  Supports full path and relative path in report.xml file.
+   */     
+  private String getCaseSensitiveFileName(String file, List<java.io.File> sourceDirs) {
+    // check whether the report file uses absolute path
+    File targetfile = new java.io.File(file);
+    if (targetfile.exists()) {
+      file = getRealFileName(targetfile);
+    } else {
+      Iterator<java.io.File> iterator = sourceDirs.iterator();
+      while (iterator.hasNext()) {              
+           targetfile = new java.io.File(iterator.next().getPath() + java.io.File.pathSeparator + file);
+           if (targetfile.exists()) {
+               file = getRealFileName(targetfile);
+               break;
+           }
+      }
+    }
+    return file;      
+  }
+     
+  /**
+   * Find the case sensitive file name - tools might use different naming schema
+   * e.g. VC HTML or build log report uses case insensitive file name (lower case on windows)
+   */      
+  private String getRealFileName( File filename){
+     try {
+         return filename.getCanonicalFile().getAbsolutePath();
+     } catch (java.io.IOException e) {
+       CxxUtils.LOG.error("SaveViolation GetRealFileName failed '{}'", e.toString());
+       }
+     return filename.getName();
+  }
+
+  
 }
