+++ sonar-cxx-plugin/src/main/java/org/sonar/plugins/cxx/utils/CxxReportSensor.java
@@ -33,7 +33,7 @@ import org.sonar.plugins.cxx.CxxLanguage;
 
 import java.io.File;
 import java.util.ArrayList;
-import java.util.Iterator;
+import java.util.HashSet;
 import java.util.List;
 
 /**
@@ -42,6 +42,7 @@ import java.util.List;
 public abstract class CxxReportSensor implements Sensor {
   private RuleFinder ruleFinder;
   protected Settings conf = null;
+  private HashSet<String> uniqueFileName = new HashSet<String>();
   
   public CxxReportSensor() {
   }
@@ -153,7 +154,6 @@ public abstract class CxxReportSensor implements Sensor {
       Violation violation = null;
       // handles file="" situation
       if ((file != null) && (file.length() > 0)){
-        file = getCaseSensitiveFileName(file, project.getFileSystem().getSourceDirs());
         org.sonar.api.resources.File resource =
           org.sonar.api.resources.File.fromIOFile(new File(file), project);
         if (context.getResource(resource) != null) {
@@ -171,7 +171,9 @@ public abstract class CxxReportSensor implements Sensor {
             }
           }
         } else {
-          CxxUtils.LOG.warn("Cannot find the file '{}', skipping violation '{}'", file, msg);
+          if (uniqueFileName.add(file)) {
+          CxxUtils.LOG.warn("Cannot find the file '{}', skipping violations", file);
+          }
         }
       } else {
         // project level violation
@@ -185,40 +187,6 @@ public abstract class CxxReportSensor implements Sensor {
     } else {
       CxxUtils.LOG.warn("Cannot find the rule {}, skipping violation", ruleId);
     }
-  }
-  
-  /**
-   *  Supports full path and relative path in report.xml file.
-   */     
-  private String getCaseSensitiveFileName(String file, List<java.io.File> sourceDirs) {
-    // check whether the report file uses absolute path
-    File targetfile = new java.io.File(file);
-    if (targetfile.exists()) {
-      file = getRealFileName(targetfile);
-    } else {
-      Iterator<java.io.File> iterator = sourceDirs.iterator();
-      while (iterator.hasNext()) {              
-           targetfile = new java.io.File(iterator.next().getPath() + java.io.File.pathSeparator + file);
-           if (targetfile.exists()) {
-               file = getRealFileName(targetfile);
-               break;
-           }
-      }
-    }
-    return file;      
-  }
-     
-  /**
-   * Find the case sensitive file name - tools might use different naming schema
-   * e.g. VC HTML or build log report uses case insensitive file name (lower case on windows)
-   */      
-  private String getRealFileName( File filename){
-     try {
-         return filename.getCanonicalFile().getAbsolutePath();
-     } catch (java.io.IOException e) {
-       CxxUtils.LOG.error("SaveViolation GetRealFileName failed '{}'", e.toString());
-       }
-     return filename.getName();
   }
   
   protected void processReport(Project project, SensorContext context, File report)
